Revert "Refacto `SearchContext` et mise en place de l'api de calcul de créneaux pour RDVI (#3827)"
This reverts commit 7589d69c33f57d08cfb2b1b292f484c4e6b7e2f6.
