+++ .rubocop.yml
@@ -45,7 +45,6 @@ Metrics/BlockLength:
     - "config/initializers/*"
     - "config/environments/*"
     - "config/routes.rb"
-    - "config/routes/api.rb"
 
 Metrics/CyclomaticComplexity:
   Max: 8 # default is 7
@@ -153,7 +152,6 @@ RSpec/MultipleDescribes:
 RSpec/VariableName:
   Exclude:
     - "spec/requests/api/v1/**/*"
-    - "spec/requests/api/rdvinsertion/**/*"
     - "spec/support/api_spec_shared_examples.rb"
 
 RSpec/MultipleExpectations:
