+++ spec/controllers/search_controller_spec.rb
@@ -99,7 +99,37 @@ RSpec.describe SearchController, type: :controller do
         expect(subject).not_to include("RSA orientation 3")
       end
 
+      context "when a motif category is passed" do
+        it "retrieves motifs from the selected category" do
+          get :search_rdv, params: {
+            address: address, departement: departement_number, city_code: city_code, motif_category_short_name: "rsa_orientation",
+          }
+          expect(subject).to include("RSA orientation 1")
+          expect(subject).not_to include("RSA orientation 2")
+        end
+      end
+
+      context "when there are seach terms" do
+        let!(:geo_search) do
+          instance_double(Users::GeoSearch, available_motifs: Motif.where(id: [motif.id, motif2.id, motif3.id, motif4.id]))
+        end
+
+        it "lists the available motifs that match the search terms" do
+          get :search_rdv, params: {
+            address: address, departement: departement_number, city_code: city_code, motif_search_terms: "RSA orientation",
+          }
+          expect(subject).to include("RSA orientation 1")
+          expect(subject).to include("RSA orientation 2")
+          expect(subject).to include("RSA orientation 3")
+          expect(subject).not_to include("Motif numéro 4")
+        end
+      end
+
       context "for an invitation" do
+        let!(:geo_search) do
+          instance_double(Users::GeoSearch, available_motifs: Motif.where(id: [motif2.id]))
+        end
+
         before do
           request.session["invitation"] = {
             address: address, departement: departement_number, city_code: city_code, invitation_token: invitation_token,
@@ -107,21 +137,7 @@ RSpec.describe SearchController, type: :controller do
           }
         end
 
-        context "when a motif category is passed" do
-          it "retrieves motifs from the selected category" do
-            get :search_rdv, params: {
-              address: address, departement: departement_number, city_code: city_code, motif_category_short_name: "rsa_orientation",
-            }
-            expect(subject).to include("RSA orientation 1")
-            expect(subject).not_to include("RSA orientation 2")
-          end
-        end
-
         context "when there are matching motifs for the geo search available_motifs" do
-          let(:geo_search) do
-            instance_double(Users::GeoSearch, available_motifs: Motif.where(id: [motif2.id]))
-          end
-
           it "lists the available motifs" do
             get :search_rdv
             expect(subject).to include("RSA orientation 2")
@@ -140,6 +156,14 @@ RSpec.describe SearchController, type: :controller do
             expect(subject).not_to include("RSA orientation 2")
             expect(subject).not_to include("RSA orientation 3")
             expect(subject).not_to include("Motif numéro 4")
+          end
+
+          it "reveals a problem when no motifs are available" do
+            get :search_rdv, params: {
+              organisation_ids: [other_org.id], motif_search_terms: "something random",
+            }
+            expect(subject).to include("Un problème semble s'être produit pour votre invitation. Toutes nos excuses pour cela.")
+            expect(subject).to include("support@rdv-solidarites.fr")
           end
         end
       end
