+++ spec/services/search_context_spec.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-RSpec.shared_examples "SearchContext" do
+describe SearchContext, type: :service do
+  subject { described_class.new(user: user, query_params: query_params) }
+
   let!(:user) { create(:user, organisations: [organisation]) }
   let!(:organisation) { create(:organisation) }
   let!(:service) { create(:service) }
@@ -64,7 +66,56 @@ RSpec.shared_examples "SearchContext" do
       let!(:geo_search) { instance_double(Users::GeoSearch, available_motifs: Motif.where(id: [motif.id, motif2.id])) }
 
       it "is the returns the two matching motifs" do
-        expect(subject.send(:matching_motifs)).to contain_exactly(motif, motif2)
+        expect(subject.send(:matching_motifs)).to match_array([motif, motif2])
+      end
+
+      context "when one of the motif does not belong to the preselected orgs" do
+        let!(:other_org) { create(:organisation) }
+
+        before { motif2.update! organisation: other_org }
+
+        it "returns only the matching motif from the preselected orgs" do
+          expect(subject.send(:matching_motifs)).to eq([motif])
+        end
+      end
+    end
+
+    context "for an invitation" do
+      subject { described_class.new(user: user, query_params: query_params, through_invitation: true) }
+
+      before do
+        query_params[:motif_category_short_name] = "rsa_orientation"
+      end
+
+      context "when there are matching motifs for the geo search" do
+        it "is the geo maching motif" do
+          expect(subject.send(:matching_motifs)).to eq([motif])
+        end
+      end
+
+      context "when there are no matching motifs for the geo search" do
+        before do
+          allow(Motif).to receive(:available_for_booking)
+            .and_return(Motif.where(id: motif2.id))
+          query_params[:motif_category_short_name] = "rsa_orientation_on_phone_platform"
+        end
+
+        it "is the organisation matching motif" do
+          expect(subject.send(:matching_motifs)).to eq([motif2])
+        end
+      end
+
+      context "when agents are specified" do
+        before { query_params[:referent_ids] = [agent.id] }
+
+        let!(:agent) { create(:agent, users: [user]) }
+        let!(:motif) { create(:motif, follow_up: true, motif_category: rsa_orientation, organisation: organisation) }
+        let!(:plage_ouverture) { create(:plage_ouverture, agent: agent, motifs: [motif]) }
+        let!(:geo_search) { instance_double(Users::GeoSearch, available_motifs: Motif.where(id: [motif.id, motif2.id])) }
+
+        it "is the motifs related to agent" do
+          expect(subject.send(:matching_motifs)).to eq([motif])
+        end
       end
     end
   end
@@ -83,7 +134,7 @@ RSpec.shared_examples "SearchContext" do
   end
 
   describe "#service" do
-    it "returns service from service_id params when given" do
+    it "returns serice from service_id params when given" do
       service = create(:service)
       search_context = described_class.new(user: nil, query_params: { service_id: service.id })
       expect(search_context.service).to eq(service)
@@ -195,6 +246,27 @@ RSpec.shared_examples "SearchContext" do
       motif = create(:motif, collectif: false)
       create(:plage_ouverture, motifs: [motif], lieu: lieu)
       expect(search_context.filter_motifs(Motif.where(id: motif.id))).to eq([motif])
+    end
+
+    it "returns motif when user is invited (and motif's bookable_by is agents_and_prescripteurs_and_invited_users)" do
+      # This test will be obsolete when the invitation behavior will be integrated in Rdv-s (https://github.com/betagouv/rdv-solidarites.fr/issues/3438)
+      lieu = create(:lieu)
+      query_params[:motif_category_short_name] = "rsa_orientation"
+      query_params[:lieu_id] = lieu.id
+      search_context = described_class.new(user: nil, query_params: query_params, through_invitation: true)
+      motif = create(:motif, bookable_by: :agents_and_prescripteurs_and_invited_users, motif_category: rsa_orientation, organisation: organisation)
+      create(:plage_ouverture, motifs: [motif], lieu: lieu)
+      expect(search_context.filter_motifs(Motif.where(id: motif.id))).to eq([motif])
+    end
+
+    it "doesnt returns motif when user is not invited (and motif's bookable_by is agents_and_prescripteurs_and_invited_users)" do
+      lieu = create(:lieu)
+      query_params[:motif_category_short_name] = "rsa_orientation"
+      query_params[:lieu_id] = lieu.id
+      search_context = described_class.new(user: nil, query_params: query_params)
+      motif = create(:motif, bookable_by: :agents_and_prescripteurs_and_invited_users, motif_category: rsa_orientation, organisation: organisation)
+      create(:plage_ouverture, motifs: [motif], lieu: lieu)
+      expect(search_context.filter_motifs(Motif.where(id: motif.id))).to eq([])
     end
   end
 end
